extract commp and dagify to lib
